chore: refactor Langchain unit tests
