+++ api/terraform/python/lambda_langchain/tests/init.py
@@ -1,0 +1,7 @@
+import json
+
+
+def get_event(filespec):
+    with open(filespec, "r") as f:
+        event = json.load(f)
+        return event
